--- a/src/daft-core/src/lib.rs
+++ b/src/daft-core/src/lib.rs
@@ -5,6 +5,7 @@ pub mod count_mode;
 pub mod datatypes;
 #[cfg(feature = "python")]
 pub mod ffi;
+pub mod join;
 pub mod kernels;
 #[cfg(feature = "python")]
 pub mod python;
@@ -16,6 +17,7 @@ use pyo3::prelude::*;
 
 pub use count_mode::CountMode;
 pub use datatypes::DataType;
+pub use join::{JoinStrategy, JoinType};
 pub use series::{IntoSeries, Series};
 
 pub const VERSION: &str = env!("CARGO_PKG_VERSION");
@@ -31,6 +33,8 @@ pub const DAFT_BUILD_TYPE: &str = {
 #[cfg(feature = "python")]
 pub fn register_modules(_py: Python, parent: &PyModule) -> PyResult<()> {
     parent.add_class::<CountMode>()?;
+    parent.add_class::<JoinType>()?;
+    parent.add_class::<JoinStrategy>()?;
 
     Ok(())
 }