--- a/src/daft-micropartition/src/ops/join.rs
+++ b/src/daft-micropartition/src/ops/join.rs
@@ -1,5 +1,5 @@
 use common_error::DaftResult;
-use daft_core::array::ops::DaftCompare;
+use daft_core::{array::ops::DaftCompare, join::JoinType};
 use daft_dsl::ExprRef;
 use daft_io::IOStatsContext;
 use daft_table::infer_join_schema;
@@ -14,36 +14,43 @@ impl MicroPartition {
         right: &Self,
         left_on: &[ExprRef],
         right_on: &[ExprRef],
+        how: JoinType,
     ) -> DaftResult<Self> {
         let io_stats = IOStatsContext::new("MicroPartition::hash_join");
         let join_schema = infer_join_schema(&self.schema, &right.schema, left_on, right_on)?;
 
-        if self.len() == 0 || right.len() == 0 {
+        if (how == JoinType::Inner && (self.len() == 0 || right.len() == 0))
+            || (how == JoinType::Left && self.len() == 0)
+            || (how == JoinType::Right && right.len() == 0)
+            || (how == JoinType::Outer && (self.len() == 0 && right.len() == 0))
+        {
             return Ok(Self::empty(Some(join_schema.into())));
         }
 
-        let tv = match (&self.statistics, &right.statistics) {
-            (_, None) => TruthValue::Maybe,
-            (None, _) => TruthValue::Maybe,
-            (Some(l), Some(r)) => {
-                let l_eval_stats = l.eval_expression_list(left_on, &self.schema)?;
-                let r_eval_stats = r.eval_expression_list(right_on, &right.schema)?;
-                let mut curr_tv = TruthValue::Maybe;
-                for (lc, rc) in l_eval_stats
-                    .columns
-                    .values()
-                    .zip(r_eval_stats.columns.values())
-                {
-                    if let TruthValue::False = lc.equal(rc)?.to_truth_value() {
-                        curr_tv = TruthValue::False;
-                        break;
+        if how == JoinType::Inner {
+            let tv = match (&self.statistics, &right.statistics) {
+                (_, None) => TruthValue::Maybe,
+                (None, _) => TruthValue::Maybe,
+                (Some(l), Some(r)) => {
+                    let l_eval_stats = l.eval_expression_list(left_on, &self.schema)?;
+                    let r_eval_stats = r.eval_expression_list(right_on, &right.schema)?;
+                    let mut curr_tv = TruthValue::Maybe;
+                    for (lc, rc) in l_eval_stats
+                        .columns
+                        .values()
+                        .zip(r_eval_stats.columns.values())
+                    {
+                        if let TruthValue::False = lc.equal(rc)?.to_truth_value() {
+                            curr_tv = TruthValue::False;
+                            break;
+                        }
                     }
+                    curr_tv
                 }
-                curr_tv
+            };
+            if let TruthValue::False = tv {
+                return Ok(Self::empty(Some(join_schema.into())));
             }
-        };
-        if let TruthValue::False = tv {
-            return Ok(Self::empty(Some(join_schema.into())));
         }
 
         // TODO(Clark): Elide concatenations where possible by doing a chunk-aware local table join.
@@ -53,7 +60,7 @@ impl MicroPartition {
         match (lt.as_slice(), rt.as_slice()) {
             ([], _) | (_, []) => Ok(Self::empty(Some(join_schema.into()))),
             ([lt], [rt]) => {
-                let joined_table = lt.hash_join(rt, left_on, right_on)?;
+                let joined_table = lt.hash_join(rt, left_on, right_on, how)?;
                 Ok(MicroPartition::new_loaded(
                     join_schema.into(),
                     vec![joined_table].into(),