--- a/src/daft-plan/src/lib.rs
+++ b/src/daft-plan/src/lib.rs
@@ -4,7 +4,6 @@
 
 mod builder;
 mod display;
-mod join;
 mod logical_ops;
 mod logical_optimization;
 mod logical_plan;
@@ -21,7 +20,6 @@ mod treenode;
 
 pub use builder::{LogicalPlanBuilder, PyLogicalPlanBuilder};
 use daft_scan::file_format::FileFormat;
-pub use join::{JoinStrategy, JoinType};
 pub use logical_plan::LogicalPlan;
 pub use partitioning::ClusteringSpec;
 pub use physical_plan::PhysicalPlanScheduler;
@@ -49,8 +47,6 @@ pub fn register_modules(_py: Python, parent: &PyModule) -> PyResult<()> {
     parent.add_class::<JsonSourceConfig>()?;
     parent.add_class::<CsvSourceConfig>()?;
     parent.add_class::<DatabaseSourceConfig>()?;
-    parent.add_class::<JoinType>()?;
-    parent.add_class::<JoinStrategy>()?;
     parent.add_class::<PhysicalPlanScheduler>()?;
     parent.add_class::<ResourceRequest>()?;
     parent.add_class::<FileInfos>()?;