Currently does not support using sort-merge joins, but broadcast and hash joins will work fine